+++ backend/src/data/mod.rs
@@ -21,7 +21,7 @@ pub struct SightsConfig {
 pub struct CategoryTagMap {
     category: String,
     opening_hours: String,
-    duration_of_stay_minutes : u32,
+    duration_of_stay_minutes : usize,
     tags: Vec<Tag>
 }
